+++ saleor/graphql/order/sorters.py
@@ -1,7 +1,7 @@
 from django.db.models import CharField, ExpressionWrapper, OuterRef, QuerySet, Subquery
 
 from ...payment.models import Payment
-from ..core.descriptions import DEPRECATED_IN_3X_INPUT
+from ..core.descriptions import ADDED_IN_322
 from ..core.doc_category import DOC_CATEGORY_ORDERS
 from ..core.types import BaseEnum, SortInputObjectType
 
@@ -9,12 +9,13 @@ from ..core.types import BaseEnum, SortInputObjectType
 class OrderSortField(BaseEnum):
     NUMBER = ["number"]
     RANK = ["search_rank", "id"]
-    CREATION_DATE = ["created_at", "status", "pk"]
-    CREATED_AT = ["created_at", "status", "pk"]
-    LAST_MODIFIED_AT = ["updated_at", "status", "pk"]
-    CUSTOMER = ["billing_address__last_name", "billing_address__first_name", "pk"]
-    PAYMENT = ["last_charge_status", "status", "pk"]
-    FULFILLMENT_STATUS = ["status", "user_email", "pk"]
+    CREATION_DATE = ["created_at", "status", "number", "pk"]
+    CREATED_AT = ["created_at", "status", "number"]
+    LAST_MODIFIED_AT = ["updated_at", "status", "number"]
+    CUSTOMER = ["billing_address__last_name", "billing_address__first_name", "number"]
+    PAYMENT = ["last_charge_status", "status", "number"]
+    FULFILLMENT_STATUS = ["status", "user_email", "number"]
+    STATUS = ["status", "number"]
 
     class Meta:
         doc_category = DOC_CATEGORY_ORDERS
@@ -22,22 +23,31 @@ class OrderSortField(BaseEnum):
     @property
     def description(self):
         descriptions = {
+            OrderSortField.NUMBER.name: "Sort orders by number.",  # type: ignore[attr-defined] # graphene.Enum is not typed # noqa: E501
             OrderSortField.RANK.name: (  # type: ignore[attr-defined] # graphene.Enum is not typed # noqa: E501
-                "rank. Note: This option is available only with the `search` filter."
+                "Sort orders by rank. Note: This option is available only with the `search` filter."
             ),
-            OrderSortField.CREATION_DATE.name: (  # type: ignore[attr-defined] # graphene.Enum is not typed # noqa: E501
-                f"creation date. {DEPRECATED_IN_3X_INPUT}"
-            ),
+            OrderSortField.CREATION_DATE.name: "Sort orders by creation date",  # type: ignore[attr-defined] # graphene.Enum is not typed # noqa: E501
+            OrderSortField.CREATED_AT.name: "Sort orders by creation date.",  # type: ignore[attr-defined] # graphene.Enum is not typed # noqa: E501
+            OrderSortField.LAST_MODIFIED_AT.name: "Sort orders by last modified date.",  # type: ignore[attr-defined] # graphene.Enum is not typed # noqa: E501
+            OrderSortField.CUSTOMER.name: "Sort orders by customer.",  # type: ignore[attr-defined] # graphene.Enum is not typed # noqa: E501
+            OrderSortField.PAYMENT.name: "Sort orders by payment status.",  # type: ignore[attr-defined] # graphene.Enum is not typed # noqa: E501
+            OrderSortField.FULFILLMENT_STATUS.name: "Sort orders by fulfillment status.",  # type: ignore[attr-defined] # graphene.Enum is not typed # noqa: E501
+            OrderSortField.STATUS.name: f"Sort orders by order status. {ADDED_IN_322}",  # type: ignore[attr-defined] # graphene.Enum is not typed # noqa: E501
         }
-
-        if self.name in OrderSortField.__enum__._member_names_:
-            if self.name in descriptions:
-                return f"Sort orders by {descriptions[self.name]}"
-
-            sort_name = self.name.lower().replace("_", " ")
-            return f"Sort orders by {sort_name}."
-
+        if self.name in descriptions:
+            return descriptions[self.name]
         raise ValueError(f"Unsupported enum value: {self.value}")
+
+    @property
+    def deprecation_reason(self):
+        deprecations = {
+            OrderSortField.CREATION_DATE.name: ("Use `createdAt` instead."),  # type: ignore[attr-defined] # graphene.Enum is not typed # noqa: E501
+            OrderSortField.FULFILLMENT_STATUS.name: ("Use `status` instead."),  # type: ignore[attr-defined] # graphene.Enum is not typed # noqa: E501
+        }
+        if self.name in deprecations:
+            return deprecations[self.name]
+        return None
 
     @staticmethod
     def qs_with_payment(queryset: QuerySet, **_kwargs) -> QuerySet: